+++ app/Models/Article.php
@@ -41,4 +41,9 @@ class Article extends Model
 
         return $this->likes()->where('user_id', $user->id)->exists();
     }
+
+    public function comments()
+    {
+        return $this->hasMany(Comment::class);
+    }
 }